CLDR assets generated
