+++ packages/localization/package.json
@@ -29,6 +29,7 @@
     "escodegen": "^1.11.0",
     "esprima": "^4.0.1",
     "glob": "^7.1.3",
+    "mkdirp": "^0.5.1",
     "path-exists-cli": "^1.0.0",
     "replace-in-file": "^3.4.2",
     "rimraf": "^2.6.2"
