+++ packages/localization/src/Assets.js
@@ -1,2 +1,2 @@
 // Currently the base package provides CLDR assets only
-import "./json-imports/LocaleData.js";
+import "./generated/json-imports/LocaleData.js";
