+++ packages/tools/build-configuration.json
@@ -2,5 +2,80 @@
   "theming": {
     "defaultTheme": "sap_fiori_3",
     "optionalThemes": ["sap_fiori_3_dark", "sap_belize", "sap_belize_hcb", "sap_belize_hcw"]
-  }
+  },
+  "locales": [
+    "ar",
+    "ar_EG",
+    "ar_SA",
+    "bg",
+    "ca",
+    "cs",
+    "da",
+    "de",
+    "de_AT",
+    "de_CH",
+    "el",
+    "el_CY",
+    "en",
+    "en_AU",
+    "en_GB",
+    "en_HK",
+    "en_IE",
+    "en_IN",
+    "en_NZ",
+    "en_PG",
+    "en_SG",
+    "en_ZA",
+    "es",
+    "es_AR",
+    "es_BO",
+    "es_CL",
+    "es_CO",
+    "es_MX",
+    "es_PE",
+    "es_UY",
+    "es_VE",
+    "et",
+    "fa",
+    "fi",
+    "fr",
+    "fr_BE",
+    "fr_CA",
+    "fr_CH",
+    "fr_LU",
+    "he",
+    "hi",
+    "hr",
+    "hu",
+    "id",
+    "it",
+    "it_CH",
+    "ja",
+    "kk",
+    "ko",
+    "lt",
+    "lv",
+    "ms",
+    "nb",
+    "nl",
+    "nl_BE",
+    "pl",
+    "pt",
+    "pt_PT",
+    "ro",
+    "ru",
+    "ru_UA",
+    "sk",
+    "sl",
+    "sr",
+    "sv",
+    "th",
+    "tr",
+    "uk",
+    "vi",
+    "zh_CN",
+    "zh_HK",
+    "zh_SG",
+    "zh_TW"
+  ]
 }
